ms_embedding_documents/requirements.txt
azure-search-documents==11.5.2
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain_openai==0.2.12
langchain_text_splitters==0.3.2